Candidate Gene Pipeline code
This code takes in mutation tab delimited files and has parameters to read in other types of database files. The code gathers information from the database files and adds it to the mutation line in the input file and writes a file with the gathered information for each mutation in the file.